Add neutron-odl-ovs-hostconfig binary only on dlrn builds
The upstram networking-odl project does not have a stable/newton
branch yet, so we are currently packaging the 2.0.0 release, which
belongs to mitaka. One binary (neutron-odl-ovs-hostconfig) is
missing on that release but present on master, breaking DLRN builds.

Using the dlrn macro to temporarily add it only on DLRN builds but
not on CBS. Once a proper stable/newton release is tagged, we will
be able to remove the conditional.

Change-Id: I981a7ff75523a7fced2f185cabd84218f88f8049